nibb: Implement internal build functionality
When building internally, use the local layer listings and set the
INHERIT variable to define the "own-mirrors" behavior.

These changes ensure that we're pulling from the NI mirrors
and sources (and preferring those to the upstream) when building
with INTERNAL_BUILD set. This improves reliability while using
the same machinery to build interally and externally

Signed-off-by: Brad Mouring <brad.mouring@ni.com>