nibb: Fix "update" function
Due to some differences with how the submodules work on a
system that is not the system that setup the submodules, the
update mechanism needed to be reworked.

Basically, we want to init (problem 1) to make sure that the
master project is aware of the submodules. After that, we want
to use the submodules functionality itself to update the
submodules (problem 2) since the submodules, when configured from
a fresh system, have no notion of the branch that they are following,
this is kept in the superproject and therefore should be triggered
from that context.

Signed-off-by: Brad Mouring <brad.mouring@ni.com>
Reported-by: Haris Okanovic <haris.okanovic@ni.com>