nibb.sh: don't expand PATH in env file
Expanding ${PATH} in nibb.sh makes the PATH setting much longer than it
needs to be, and also effects a latching of PATH settings at nibb.sh
execution time -- so if a developer installs/uninstalls a package on
their machine which affects PATH, the PATH setting would get reverted
when executing the environment file. Fix by escaping the '$'.

Signed-off-by: Richard Tollerton <rich.tollerton@ni.com>
Acked-by: Brad Mouring <brad.mouring@ni.com>