.gitmodules: equalize URLs with review-board
The repos as configured in review-board universally point to
git://git.amer.corp.natinst.com, so in order for post-review to "just
work", the URLs need to be set to that, not git://git.natinst.com/.

Additionally, several repo URLs were set to git@git, which only works
for developers with pubkey access to the server. These need to change to
git://git.amer.corp.natinst.com too.

Signed-off-by: Richard Tollerton <rich.tollerton@ni.com>