nibb.sh: prune BB_ENV_EXTRAWHITE
The overwhelming majority of the variables in BB_ENV_EXTRAWHITE are not
actually set. A large number were inherited from VirtualBench's build
process and were never used by the nilrt build process. TCMODE and
TCLIBC are needed if you're building with codesourcery, which we're not.
etc.

Prune this down to what we're actually using or have some expectation of
using.

Signed-off-by: Richard Tollerton <rich.tollerton@ni.com>