.gitignore: refactor
Instead of ignoring all of build/ and un-ignoring individual files
inside it, ignore only those files we know to vary on local machines,

Subdirectories of build/, like cache/ and download/, ought to be ignored
specifically inside build/, and not ignored across the entire repo.

AFAIK export/ is not a part of the nominal build process so stop
ignoring it.

Editor autosave files like *.swp are generally useful to see in `git
status`, insofar as they are a useful and independent view of unsaved
changes.

Signed-off-by: Richard Tollerton <rich.tollerton@ni.com>