README: Update for Comms-2.0
The instructions would direct users to update submodules using
"... --remote --checkout", and that would result in versions of the
submodules that don't match what was used to build the Comms-2.0 image.

Additionally, the instructions referenced Zynq which was not supported
nor released for Comms-2.0, so remove those references to decrease
confusion.

Signed-off-by: Brad Mouring <brad.mouring@ni.com>