migration-test.py: fix intermittent network config failures
Jenkins has seen intermittent test failures because, in certain
situations, the qemu VM doesn't get its IP address as a result of the
automatic call to dhcp client at boot. I'm asuming this is a bug
somewhere either in qemu's virtualized NIC or in its usermode
networking stack (this is used by default in the test because it
doesn't need root-access modifications on the host to run VM's).

Explicitely call dhclient to correctly get an ip once the target
is booted up while waiting for the network. This is a no-op if the
target gets the ip correctly at boot in the first place.

Signed-off-by: Ioan-Adrian Ratiu <adrian.ratiu@ni.com>