ni-oe-init-build-env: Whitelist build ID/tagging vars
Allow an externally-defined variable to be used for, among other
reasons, to add to the Linux kernel version to improve visibility
into the version of a kernel being used for a particular image.

Also, provide a variable to enable the tagging and pushing of those
tags to a git server for improved traceability.

Signed-off-by: Brad Mouring <brad.mouring@ni.com>
Acked-by: Adrian Ratiu <adrian.ratiu@ni.com>
Acked-by: Ken Sharp <ken.sharp@ni.com>
Natinst-ReviewBoard-ID: 194113