Jenkinsfile: cleanup morgue
Packages being clobbered and put in the morgue is not that problematic
as I initially thought, it's fairly common when reusing sstate like we
do and updating just the package metadata which happens frequently.

Don't stop the build for this because we have a chicken and egg dilema
if we stop: the PR server is updated at the end of build but because a
morgue stops builds before updating PRs, every build stops until the
morgue is manually removed and the sstate cleared (each new package
build reuses the same PR and sends the previous package to the morgue
once the initial condition of updating just the metadata has been hit).

Signed-off-by: Ioan-Adrian Ratiu <adrian.ratiu@ni.com>