Jenkinsfile: simplify node architecture
Make use of the fact that NILRT build directories are perishable and
can be "spun up" on the fly via calls to ni-oe-init-build-env to build
the pipeline in a unified workspace for all the distro flavours (atm
x64 and arm), each in a separate build dir in a single container per
pipeline.

This cleans up the code quite a lot and makes the pipeline more
efficient (no more parallel git clones in each workspace etc).

The downside of this architecture is that we can't offload containers
on separate build machines per distro_flavour in the same pipeline.
For example for the cobra pipeline x64 on pinkman and the arm build on
boetticher in parallel. This is no big problem because we didn't
intend to run the pipeline like this in the first place (we'd be too
network bounded for the many gb/tb of data transfers between machines).

Signed-off-by: Ioan-Adrian Ratiu <adrian.ratiu@ni.com>