migration-test: cleanup & simplify for Cobra
Found a better workaround for the network bringup (a VM reboot instead
of the unreliable wait loop). There is bug report filed, CAR 685370 to
fix the underlying issue of drivers not being loaded.

Also we now have officially supported NXG VM images exported so
there's no need to provision an image in the test.

Signed-off-by: Ioan-Adrian Ratiu <adrian.ratiu@ni.com>