Jenkinsfile: move package grep outside nibuild init
For some reason the last grep cmd is not run depending on the result
of the p4 sync before it. We need the grep cmd to always run and fail
the build if it encounters an error (the export version can't be
detected) so make it run in it's own shell call ensuring it always
runs and fails the build if necessary.

Signed-off-by: Ioan-Adrian Ratiu <adrian.ratiu@ni.com>