scripts/docker: use dockerfile instead of template
Dockerfile:
* Update debian base image to 9.8 stable.
* Update Virtualbox to 6.0.
* Do not ``apt purge`` packages in the Dockerfile to avoid dumping
  needed toolchain components like guestfish.

Now that we do not need to write in a static uid/gid for the docker
build container user, we do not need scripting to templatize the
dockerfile.

Further, remove the helper scripts for building/submitting the export
because this is highly site-specific and unlikely to be useful to anyone
outside of National Instruments.

Signed-off-by: Alex Stewart <alex.stewart@ni.com>
Acked-by: Haris Okanovic <haris.okanovic@ni.com>
Natinst-ReviewBoard-ID: 281129