vpn-test.py: do not use NIAuth to login to runmode
Recent builds of the NILRT/8.0 systemlink Base Image do not include the
NIAuth PAM modules. As a result, we cannot rely on NIAuth to login to
runmode. Instead, chroot into the userfs and give the 'root' user an
arbitrary password. Use 'root' and the shadow file to authenticate login
to the runmode.

Signed-off-by: Alex Stewart <alex.stewart@ni.com>
Acked-by: Haris Okanovic <haris.okanovic@ni.com>
Natinst-ReviewBoard-ID: 309069