Change vpn test to use fixed mac addresses
With random mac addresses, the vpn test generates new leases for DHCP-generated
IP addresses on each test run. When running the test multiple times in a row,
this can quickly exhaust all available IP addresses on the subnet and cause
the test to fail. Change to fixed mac addresses to avoid this issue.

Signed-off-by: Jeffrey Pautler <jeffrey.pautler@ni.com>
Acked-by: Alex Stewart <alex.stewart@ni.com>
Natinst-ReviewBoard-ID: 313093