buildVM.sh: Add option to specify answer file on command line and quiet output
Add an option (-a) to specify an answer file on the command line.
In addition, added some quiet flags to some of the qemu/iso building
commands that add nothing to the output of the script.  Disabled the
qemu output by default and added the (-v) verbose flag to enable it
again.

Signed-off-by: Bill Pittman <bill.pittman@ni.com>
Acked-by: Haris Okanovic <haris.okanovic@ni.com>
Acked-by: Alex Stewart <alex.stewart@ni.com>
Natinst-ReviewBoard-ID: 316692