scripts/azdo: add AZDO component conf files
Add an auto.conf file, containing bitbake configuration settings which
(a) are common to all NI-internal AZDO bitbake components and (b) have
static values.

Add an ni-org.conf file, containing bitbake configuration settings which
are necessary to produce faster, more-reproducible builds, but which are
insane for an external (non-NI) builder.

Signed-off-by: Alex Stewart <alex.stewart@ni.com>