scripts: update Dockerfiles
Merge jeminor's Dockerfile, which contains the necessary modifications
to make the build image a viable NIbuild/AZDO container. But break it
into two pieces: under scripts/docker and scripts/azdo, for the generic
and NI-specific components respectively.

The Makefile in :scripts/azdo can be used to create a new nilrt-build
and nilrt-nibuild image. The nilrt-nibuild image should be pushed to the
NI docker registry for use by the build pipeline.