fixup! test_provisioning: fixup timeout messages
Using the newline character, instead of the carriage return, is better
for `send_user` in a GNU make context. When using \n before and after
the message, MAKE will avoid its tendancy to drop the first character of
the message output and also to overwrite the message line with the next
stderr output.

Prevously, messages would sometimes look like this:
```
IMEOUT waiting for opkg to configure and ptests to install.
INFO: Feed server (2473) terminated.
```

or this
```
INFO: Feed server (2473) terminated. and ptests to install.
```

Now they are printed in their completion on thier own line.