buildVM: improve KVM detection
The buildVM script attempts to enable KVM, if the calling user is a
member of the `kvm` group. That detection isn't fully reliable in
container environments.

Instead, check to see if the user can write to /dev/kvm. If that file
does not exist or is not writable, it is reasonable to assume that the
KVM module is not loaded or the user does not have KVM permissions
(respectively.) In that case, do not `-enable-kvm`.

Also, assert that we should use the qemu64 pseudo-model - which
evaluates to a common x86_64 model, and is preferred by the QEMU docs
over `kvm64`.

Signed-off-by: Alex Stewart <alex.stewart@ni.com>