azdo: Add script to fetch/push sstate-cache
This script will be used by the azdo jobs to sync down the sstate
cache from argo at build time, as well as push the local sstate cache
to argo after the build finishes.
The script maintains a maximum of five sstate cache directories on
the remote location and will delete older directories to maintain the
the limit.

Azdo Bug: 1299743

Signed-off-by: Shruthi Ravichandran <shruthi.ravichandran@ni.com>