provisioning tests: enable multi-script expect sessions
The grub-gateway ptests would be much more efficient if we could reduce
the number of times we need to standup/teardown the VM between migrating
it back to safemode and running the ptest scripts.

Give the start-vm-using-expect script the ability to run multiple expect
scripts per session, so that they can be chained in the most efficient
way.

Signed-off-by: Alex Stewart <alex.stewart@ni.com>