init: define GIT_REPODIR
The GIT_REPODIR variable is used by newer bblayers.conf files to locate
the OE layer repos. It is normally defined as a part of the NI-internal
sumo build pipeline environment, but is undefined-by-default in external
builders' environments. As a result, bitbake will refuse to comprehend
the bblayers.conf.

Define GIT_REPODIR in the init script and pass its value to bitbake, to
enable external builders.

Signed-off-by: Alex Stewart <alex.stewart@ni.com>