test_migration_to_rauc: reinforce getty login
Now that NIAuth is available in the image, switch the login user to
'admin.

When using the admin user to login, getty prompts for a password.
Exploit the exp_continue operation to optionally handle password entry,
when a login user requires it.

Use a constant to store the shell prompt regex, so that it is easier to
maintain; and make it flexible between the 'root' and 'admin' users,
because they are equivalent on a NILRT system.

Signed-off-by: Alex Stewart <alex.stewart@ni.com>