tests/provisioning: remove opkg base-feeds during test
During the provisioning and migration test runs, the test VM uses the
same base-feeds.conf opkg sources as the canonical image. The
locally-built IPK feeds are applied atop those base feeds. Though the
test VM *should* prefer the locally-built IPKs, it is possible for a
malformed VM or test configuration to invert the priority, and a test
run might pass based upon the outdated content in the base-feeds.

To avoid false positives, add a step to the test scripts to remove the
base-feeds.conf file, so that the only valid opkg sources are the
locally-build IPK feeds.

Signed-off-by: Alex Stewart <alex.stewart@ni.com>