scripts/docker: upgrade Dockerfile for hardknott
The Dockerfile in scripts/docker provides a limited container
implementation which contains the host dependencies needed to build
NILRT.

Replace it with a more complete solution, based on Debian 10.10, which
is appropriate for building NILRT on the hardknott mainline. Also
include a docker-compose file and shell script entrypoint, which
provides "one command" entry into a viable build workspace.

Signed-off-by: Alex Stewart <alex.stewart@ni.com>