scripts: add cve-to-csv.py script
The cve_check class in OE will produce a directory of information about
what reported CVEs are outstanding in the layer stack. The CVE files are
presented in plain-text stanzas, which are not very easy to sort and
filter.

cve-to-csv.py will automatically digest the cve_check directory into a
comma separated value data structure - which it writes to stdout.

Signed-off-by: Alex Stewart <alex.stewart@ni.com>