ni-oe-init-build-env: autocalculate BB_NUMBER_THREADS
The BB_NUMBER_THREADS variable determines the maximum number of bitbake
tasks which are allowed to execute contemporaneously.

If the user asserts a value for this variable, pass it through to
bitbake. Otherwise, calculate the value by doubling the core count -
which is the recommended rule of thumb from the yocto manual.

Signed-off-by: Alex Stewart <alex.stewart@ni.com>