move docker files to root and update to use pyrex
Move the docker/ directory to the project root, since docker is becoming
a first-class feature, and I assert that putting the docker folder
front-and-center in the project root is modern.

Also update our docker setup to build off of the images that the pyrex
project provides. Include a new create-... script to generate and tag
new `build-nilrt` containers. This script can be used by both developers
and build pipelines to generate viable images.

Signed-off-by: Alex Stewart <alex.stewart@ni.com>