init: more-strongly define BUILDDIR
For the --org option to be implemented reliably, the
ni-oe-init-build-env source-file needs to be confident about the
location of the OE build/ directory, so that it may copy org-specific
conf files into it prior to invoking the OE init script.

Take a stronger stance about the location of the BUILDDIR, and define
its value early in the script, so that we know where to copy files
later.

Signed-off-by: Alex Stewart <alex.stewart@ni.com>