azdo-init: inherit build image name/tag from ENV
The AZDO build pipeline is responsible for building the docker container
which pyrex will use to build OE. When it does, it wants to tag that
image with the git hash of the nilrt.git repo, and use the image_name
defined in the AZDO component.

Modify the AZDO pipeline's pyrex.ini file to pass along the image_name
and _tag to pyrex, so that it knows which image was just built.

Signed-off-by: Alex Stewart <alex.stewart@ni.com>