build.vms: remove uuid references
Neither the vm_uuid nor disk_uuid variable is used in the newer version
of the build.vm script. Remove these variables to (a) remove the unused
code and (b) obsolete the requirement that the build host machine have
an instance of `uuidgen` - which is not a standard package in most
distros.

Signed-off-by: Alex Stewart <alex.stewart@ni.com>