build.extra-feeds: ignore error codes from the -extra feed
`bitbake --continue` will still return a nonzero error code if there are
bitbake errors, which cause the script as a whole to bail because of the
`set -e` option. This isn't desirable, because we accept as a matter of
policy that some packages in the -extra feed may fail.

Squash return codes from the -extra feed build command.

Signed-off-by: Alex Stewart <alex.stewart@ni.com>