pyrex: enable containers-by-digest
The current pyrex implementation assumes that users wish to start
containers using the `$name:$tag` format. Users may not use the
`$name@$digest` format, which is preferable in environments where
build-reproducibility is required.

Change the pyrex.ini to eschew this assumption and update the init
script to allow the user to pull containers by digest.

Signed-off-by: Alex Stewart <alex.stewart@ni.com>