ni-oe-ini-build-env: source container tag from meta-nilrt
When the user does not define a PYREX_BUILD_NILRT_IMAGE string, the init
script will default to using `build-nilrt:latest` - which is not
generally safe to use on all mainline branches of nilrt.git.

Instead, source the release codename from the meta-nilrt layer, and use
its value as the container tag.

Signed-off-by: Alex Stewart <alex.stewart@ni.com>