create-build-nilrt.sh: tag image with codename
The create-build-nilrt.sh script tags the resulting `build-nilrt`
image as 'latest', in all cases. This is dangerous on systems which also
build the sumo-mainline, as the user may inadvertently source into the
wrong container image.

Instead, source the NILRT codename from the meta-nilrt layer, and tag
the image with the codename.

Signed-off-by: Alex Stewart <alex.stewart@ni.com>