emit-feed-changelog-md.py: add script for JSON to Markdown
We can generate a diff between core/ and extra/ feeds, but we want to
display some of that info in a human-readable format. Specifically,
we want to display the packages added and removed between releases,
and that's what this script does.

Given lines of JSON diff to the script, it generates a report in
Markdown intended to be written to a file in /docs

Signed-off-by: Dylan Turner <dylan.turner@ni.com>