upstream_merge: new arg disables submod sanity
When performing an upstream merge on a non-default release, one must
perform a git checkout on each submodule as the local branches are not
auto-grabbed; they are simply set to a remote commit.

This adds optional functionality to upstream_merge.sh which forces a
git checkout on the provided local branch from repos.conf by skipping
the sanity check and continuing on to the rest of the operations.

Default behavior of doing the sanity check is preserved

Signed-off-by: Dylan Turner <dylan.turner@ni.com>