cve-to-csv.py: Handle multiline fields in CVE files.
When including cve-check, CVE entries are pulled from an
online database which we don't control. Some entries in
the database recently had new lines in the description
which caused the script to convert all the CVE files to
a CSV file to fail.

This change updates the script to parse the CVE files
using regexes instead of line by line.

Signed-off-by: Charlie Johnston <charlie.johnston@ni.com>