diff-cve-csvs.py: set sane verbosity default
The argparse module sets the default value of a 'count' action argument
to NoneType, instead of zero. This causes the script to throw a
comparison error when later evaluating the verbosity level, in cases
where the user hasn't given the '-v' argument.

Set the default to `0`, so that comparison works as expected.

Signed-off-by: Alex Stewart <alex.stewart@ni.com>