remove feed-changelogs
The feed changelogs are semi-automated artifacts that record package
feed differences between NILRT releases, in both JSON and markdown
formats. They were speculatively added on the hope that they would help
to explain how packages enter or leave the ecosystem.

In retrospect, they have not proven to be useful. And it is difficult for
the release team to remember to update them.

Deprecate these artifacts, until we have a better usecase for their
inclusion.

Signed-off-by: Alex Stewart <astewart.49c6@gmail.com>