runQEMUVM.sh: add forward_port argument
It is common to want to forward a QEMU host port to the guest, for the
purposes of SSHing into the guest VM from the host. This is especially
useful when the guest is using a networking profile that makes it
impossible to 'see' from the host perspective.

Add a command line argument to forward a host port to the guest's SSH
port.

Signed-off-by: Alex Stewart <alex.stewart@ni.com>